refactor(console): move gio-connector-dialog into api components
