+++ gravitee-apim-console-webui/src/management/api/component/gio-connector-dialog/gio-connector-dialog.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 .gio-connector-dialog {
   &__title {
